+++ src/projects/canvai-ui-system/v3/components/index.ts
@@ -1,0 +1,5 @@
+export { Swatch } from './Swatch'
+export { ScaleRow } from './ScaleRow'
+export { PreviewTopBar } from './PreviewTopBar'
+export { PreviewSidebar } from './PreviewSidebar'
+export { PreviewProjectPicker } from './PreviewProjectPicker'